add bootstrap

+++ db/migrate/20130913152801_create_tweets.rb
@@ -2,12 +2,15 @@ class CreateTweets < ActiveRecord::Migration
   def change
     create_table :tweets do |t|
       t.string      :text
-      t.integer     :tweet_rating
+      t.float     :tweet_rating
       t.datetime    :date_tweeted
-      t.integer     :tweet_id
+      t.float     :tweet_id
       t.timestamps 
 
       t.belongs_to  :politician
     end
   end
 end
+
+
+
